requirement.txt
# Core Libraries
streamlit
python-dotenv
pandas

# Document Parsing
PyPDF2
python-docx
pikepdf
chardet

# Text Processing
nltk

# Advanced AI and Embeddings
langchain-community
langchain-groq
sentence-transformers

# Vector Store
chromadb

# Development and Debugging
pytest
black
flake8